+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 
 ## [Unreleased](https://github.com/mutms/moodle-tool_muprog/compare/v5.0.6.01...HEAD)
 
-- No changes
+### Changes
+
+- Geopattern generated program images are served via standard pluginfile.php 
 
 ## [v5.0.6.01](https://github.com/mutms/moodle-tool_muprog/compare/mu-5.0.5-01...v5.0.6.01) - 2026-02-12